# swagger.yaml
openapi: 3.0.3
info:
  title: Dating App API
  version: 1.0.0
  description: |
    Music-based dating application backend. Matches users based on Spotify listening history,
    lifestyle compatibility, and behavioral learning.

    ## Authentication
    Most endpoints require a JWT passed as `Authorization: Bearer <token>`.
    JWT claims: `{ userId, tokenVersion, sub, iat, exp }`.

    After a password reset `tokenVersion` is incremented — all prior tokens are invalidated.
    Deleted-user JWTs return **401** (not 404).

    ## Email Verification
    Unverified-email users receive **403** `{"error":"Email verification required"}` on all
    endpoints except `/api/v1/auth/**`, `/api/v1/onboarding/**`, `/public/**`, and `/api/test/**`.

    ## Rate Limiting
    | Action | Limit | Key |
    |--------|-------|-----|
    | Swipe | 60/min | userId |
    | Score calculation | 30/min | userId |
    | Unauthenticated | per IP | remoteAddr |

    Exceeded limits return **429**.

    ## Error Contract
    Every error response follows:
    ```json
    { "code": "string", "message": "string", "fields": {}, "timestamp": "ISO8601" }
    ```
    Always switch on `code`, never on `message`.

  contact:
    name: API Support
    email: mladen.hangi@kodelab.hr
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/matching/potential:
    get:
      tags: [Matching]
      summary: Get paginated match recommendations
      operationId: getPotentialMatches
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Pagination offset
        - name: minScore
          in: query
          schema:
            type: number
            format: double
            minimum: 0
            maximum: 100
            default: 0
          description: Minimum compatibility score filter
        - name: excludeSwiped
          in: query
          schema:
            type: boolean
            default: true
          description: Exclude already-swiped users
      responses:
        '200':
          description: Paginated potential matches
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PotentialMatchPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/matching/score/{otherUserId}:
    get:
      tags: [Matching]
      summary: Get detailed compatibility score with another user
      description: |
        Returns the directional score (current user → other user). Rate limited to 30/min.
        Blocked relationships return 403. Self-scoring returns 400.
      operationId: getMatchScore
      parameters:
        - name: otherUserId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440001
      responses:
        '200':
          description: Detailed compatibility score
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchScore'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/matching/swipe:
    post:
      tags: [Matching]
      summary: Record a swipe action
      description: |
        Rate limited to 60/min.
        - `super_like` triggers instant match.
        - `block` prevents all future interactions and auto-unmatches any active match.
      operationId: recordSwipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwipeRequest'
            example:
              swipedUserId: 550e8400-e29b-41d4-a716-446655440001
              action: like
              platform: web
      responses:
        '200':
          description: Swipe recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwipeResult'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/DuplicateSwipe'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /api/v1/matching/matches:
    get:
      tags: [Matching]
      summary: Get paginated matches for the current user
      operationId: getMatches
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, all]
            default: active
          description: Filter by match status
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Paginated match list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchPageResponse'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/matching/analytics:
    get:
      tags: [Matching]
      summary: Get swipe analytics for the current user
      operationId: getAnalytics
      responses:
        '200':
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ─── Genre Preferences ─────────────────────────────────────────────────────
...
  /api/v1/spotify/search/artists:
    get:
      tags: [Spotify (Public)]
      summary: Search for artists
      operationId: searchArtists
      security: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
          example: Coldplay
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Artist search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistSearchResult'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/spotify/search/tracks:
    get:
      tags: [Spotify (Public)]
      summary: Search for tracks
      operationId: searchTracks
      security: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
          example: Bohemian Rhapsody
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Track search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackSearchResult'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/spotify/genres:
    get:
      tags: [Spotify (Public)]
      summary: Get all available Spotify genre seeds
      operationId: getAvailableGenres
      security: []
      responses:
        '200':
          description: List of available genre strings
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example: [acoustic, afrobeat, alt-rock, alternative, ambient, blues, classical, country, dance, electronic, folk, hip-hop, indie, jazz, metal, pop, punk, r-n-b, reggae, rock]
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/spotify/artists/{artistId}:
    get:
      tags: [Spotify (Public)]
      summary: Get artist details by Spotify ID
      operationId: getArtistById
      security: []
      parameters:
        - name: artistId
          in: path
          required: true
          schema:
            type: string
          example: 4gzpq5DuGXgMSoJzLEbfR
      responses:
        '200':
          description: Artist details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpotifyArtist'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

# ─── Components ──────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT with claims: `{ userId, tokenVersion, sub, iat, exp }`.
        `tokenVersion` is required — tokens without it are rejected with 401.

  # ─── Responses ─────────────────────────────────────────────────────────────
  responses:
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: Validation failed
            fields:
              email: must not be blank
              password: Password must contain at least one uppercase letter, one lowercase letter, one number, and one special character
            timestamp: '2026-05-12T10:15:30.123Z'
    InvalidArgument:
      description: Invalid query parameter or illegal request state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INVALID_ARGUMENT
            message: Invalid status parameter. Allowed values active, all
            timestamp: '2026-05-12T10:15:30.123Z'
...
    Unauthorized:
      description: Missing or invalid JWT, or wrong credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Invalid email or password
            timestamp: '2026-05-12T10:15:30.123Z'
    Forbidden:
      description: Authenticated but not authorized for this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: Access denied
            timestamp: '2026-05-12T10:15:30.123Z'
...
    DuplicateSwipe:
      description: User has already swiped on this person
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: DUPLICATE_SWIPE
            message: You have already swiped on this user
            timestamp: '2026-05-12T10:15:30.123Z'
...
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: User not found
            timestamp: '2026-05-12T10:15:30.123Z'
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: ACCOUNT_LOCKED
            message: Too many requests. Please slow down.
            timestamp: '2026-05-12T10:15:30.123Z'
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INTERNAL_ERROR
            message: An unexpected error occurred
            timestamp: '2026-05-12T10:15:30.123Z'

  # ─── Schemas ───────────────────────────────────────────────────────────────
  schemas:
...
    CompatibilityLevel:
      type: string
      enum: [LOW, MEDIUM, HIGH, VERY_HIGH]
      description: LOW=0-40, MEDIUM=41-60, HIGH=61-80, VERY_HIGH=81-100

    # ── Error ──────────────────────────────────────────────────────────────
...
    PotentialMatchPage:
      type: object
      properties:
        matches:
          type: array
          items:
            $ref: '#/components/schemas/PotentialMatch'
        total:
          type: integer
          example: 42
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
        hasMore:
          type: boolean
          example: true
...
    MatchScore:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        otherUserId:
          type: string
          format: uuid
        overallScore:
          type: number
          format: double
          description: 0–100
          example: 78.4
        musicScore:
          type: number
          format: double
          description: 0–100
          example: 85.0
        lifestyleScore:
          type: number
          format: double
          description: 0–100
          example: 72.0
        interestsScore:
          type: number
          format: double
          description: 0–100
          example: 65.0
        locationScore:
          type: number
          format: double
          description: 0–100
          example: 90.0
        behavioralScore:
          type: number
          format: double
          nullable: true
          description: 0–100; null during cold start (fewer than 5 likes)
          example: 70.0
        breakdown:
          $ref: '#/components/schemas/MatchBreakdown'
        insights:
          type: array
          items:
            type: string
          example:
            - You share 3 genres including indie-rock and alternative
            - Both value concerts highly
        compatibilityLevel:
          $ref: '#/components/schemas/CompatibilityLevel'
        calculatedAt:
          type: string
          format: date-time
          example: '2026-05-12T10:15:30.000Z'
    SwipeRequest:
      type: object
      required: [swipedUserId, action]
      properties:
        swipedUserId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440001
        action:
          type: string
          enum: [like, pass, super_like, block]
          example: like
        matchScore:
          type: number
          format: double
          nullable: true
          deprecated: true
          description: Deprecated — server computes the score internally; field is silently ignored
        platform:
          type: string
          maxLength: 50
          nullable: true
          enum: [web, ios, android]
          default: web
    SwipeResult:
      type: object
      properties:
        swipeId:
          type: string
          format: uuid
        action:
          type: string
          enum: [like, pass, super_like, block]
          example: like
        matchScore:
          type: number
          format: double
          description: Server-computed score at swipe time
          example: 78.4
        resultedInMatch:
          type: boolean
          example: false
        match:
          $ref: '#/components/schemas/SwipeMatchDetails'
          nullable: true
          description: Only present when resultedInMatch=true
    SwipeMatchDetails:
      type: object
      properties:
        matchId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          description: The other user's ID
        name:
          type: string
          example: Bob Jones
        matchScore:
          type: number
          format: double
          example: 78.4
        matchedAt:
          type: string
          format: date-time
          example: '2026-05-12T10:16:00.000Z'
...
    MatchPageResponse:
      type: object
      properties:
        matches:
          type: array
          items:
            $ref: '#/components/schemas/MatchResponseDto'
        total:
          type: integer
          format: int64
          example: 7
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
        hasMore:
          type: boolean
          example: false
    AnalyticsResponse:
      type: object
      properties:
        totalSwipes:
          type: integer
          format: int64
          example: 150
        totalLikes:
          type: integer
          format: int64
          example: 90
        totalPasses:
          type: integer
          format: int64
          example: 60
        totalMatches:
          type: integer
          format: int64
          example: 12
        swipeThroughRate:
          type: number
          format: double
          nullable: true
          description: likes / totalSwipes; null if no swipes
          example: 0.60
        matchRate:
          type: number
          format: double
          nullable: true
          description: matches / likes; null if no likes
          example: 0.13

    # ── Genre Preferences ──────────────────────────────────────────────────
...
    SpotifyArtist:
      type: object
      properties:
        id:
          type: string
          example: 4gzpq5DuGXgMSoJzLEbfR
        name:
          type: string
          example: Radiohead
        genres:
          type: array
          items:
            type: string
          example: [alternative rock, art rock]
        popularity:
          type: integer
          minimum: 0
          maximum: 100
          example: 78
        images:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
                format: uri
              height:
                type: integer
              width:
                type: integer
        external_urls:
          type: object
          properties:
            spotify:
              type: string
              format: uri
        followers:
          type: object
          properties:
            total:
              type: integer
...
    ArtistSearchResult:
      type: object
      description: Paginated artist search results
      properties:
        artists:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/SpotifyArtist'
            total:
              type: integer
            limit:
              type: integer
            offset:
              type: integer
            next:
              type: string
              format: uri
              nullable: true
            previous:
              type: string
              format: uri
              nullable: true
    TrackSearchResult:
      type: object
      description: Paginated track search results
      properties:
        tracks:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/SpotifyTrack'
            total:
              type: integer
            limit:
              type: integer
            offset:
              type: integer
            next:
              type: string
              format: uri
              nullable: true
            previous:
              type: string
              format: uri
              nullable: true

    # ── Utility ────────────────────────────────────────────────────────────
...